fix: RP tests some code replaced to separate fililes
